# Vectorized writing array improves its speed 